third_party/blink/web_tests/platform/generic/fast/css/pseudo-invalid-novalidate-001-expected.txt

This test checks that an input element that doesn't validate and an invalid input don't share style.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

 
PASS document.defaultView.getComputedStyle(v[0], null).getPropertyValue('background-color') is 'rgb(0, 0, 255)'
PASS document.defaultView.getComputedStyle(v[1], null).getPropertyValue('background-color') is 'rgb(0, 255, 0)'
PASS successfullyParsed is true

TEST COMPLETE
